% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_matrix.R
\name{convert_id}
\alias{convert_id}
\title{转换基因ID}
\usage{
convert_id(genes, ref, ip.type, op.type)
}
\arguments{
\item{genes}{基因向量名}

\item{ref}{参考的org对象}

\item{ip.type}{输入基因向量的类型}

\item{op.type}{输出基因向量的类型}
}
\description{
转换基因ID。仅用于一种ID的转换，返回第一个匹配的结果。
结果为一个数据框，与输入的基因向量一一对应。
}
