% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_normalize.R
\name{data_normalize}
\alias{data_normalize}
\title{计算CPM、TPM、FPKM}
\usage{
data_normalize(count_file, op.dir = ".")
}
\arguments{
\item{count_file}{featurecounts原始输出文件的绝对路径}

\item{op.dir}{保存结果文件的路径}
}
\description{
基于featurecounts算法的原始输出结果，分别计算CPM、TPM、FPKM。
}
