% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_pipeline.R
\name{deg_pipeline}
\alias{deg_pipeline}
\title{差异表达基因分析}
\usage{
deg_pipeline(counts, tpm, group, contrast, p.adj = 0.05, FC = 2, write = F)
}
\arguments{
\item{counts}{数据框。counts矩阵}

\item{tpm}{数据框。tpm矩阵}

\item{group}{样本的分组向量（最多两种分组）}

\item{contrast}{比较向量。如c('M','N')，此时以N对照，FC为M/N}

\item{p.adj}{差异表达基因的p值的阈值}

\item{FC}{差异表达基因的FC的阈值}

\item{write}{为T时，各算法的差异表达基因分析结果会自动保存至本地。}
}
\description{
自动调用DESeq2、edgeR和limma包识别差异表达基因。
}
