% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_matrix.R
\name{exp_matrix_unique_row}
\alias{exp_matrix_unique_row}
\title{处理重复的行}
\usage{
exp_matrix_unique_row(
  exp,
  raw_row,
  convert_name,
  type = "max",
  mc.cores = NULL
)
}
\arguments{
\item{exp}{表达矩阵}

\item{raw_row}{原始的基因向量（行名）}

\item{convert_name}{转换后的基因向量（与raw_row对应）。}

\item{type}{保留数据的方式。可以选max、min、mean。为mean时，返回多行数据的均值。}

\item{mc.cores}{默认调用最多线程}
}
\description{
转换ID后，可能会有多个行对应同一ID的情况，可以使用该函数进行处理，保留唯一行。
}
