% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_colors.R
\name{gg.op}
\alias{gg.op}
\title{ggplot2对象的预览和保存}
\usage{
gg.op(
  plot,
  pre = NULL,
  sur = ".jpg",
  file = NULL,
  res = 600,
  height = 5,
  width = 5
)
}
\arguments{
\item{plot}{ggplot2对象}

\item{pre}{文件名前缀}

\item{sur}{文件名后缀。默认是'.jpg'}

\item{file}{完整的文件名。调用该参数时，会忽略pre和sur}

\item{res}{DPI}

\item{height}{图片的高度}

\item{width}{图片的宽度}
}
\description{
ggplot2对象的预览和保存
}
