% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_colors.R
\name{w.ggop}
\alias{w.ggop}
\title{Previewing and saving ggplot2 objects}
\usage{
w.ggop(
  plot,
  pre = NULL,
  sur = ".png",
  file = NULL,
  res = 600,
  h = 5,
  w = 5,
  mode = "F1"
)
}
\arguments{
\item{plot}{ggplot2 object.}

\item{pre}{File name prefix.}

\item{sur}{File name suffix, default is '.png'. Supported image formats are identical to those of \code{ggpubr::ggexport}.}

\item{file}{Full file name. If this parameter is provided, pre and sur will be ignored.}

\item{res}{Dots per inch (DPI) resolution.}

\item{h}{Height of the image (inch). For raster plots, the final image height is (h × res) pixels. For vector graphics, the final height is h inches.}

\item{w}{Width of the image (inch). For raster plots, the final image width is (w × res) pixels. For vector graphics, the final width is w inches.}

\item{mode}{There are two preview modes:

(1) F1 shows a preview generated according to the specified parameters. F1 may sometimes differ slightly from the actual saved image, but it can be viewed in a separate graphics window in R.

(2) F2 shows the image that has already been saved locally, but zooming in and out is not supported.

(3) If local image preview is not required, set this parameter to FALSE.

By default, preview mode F1 is used.}
}
\description{
This function wraps and extends \code{ggpubr::ggexport} to facilitate saving ggplot objects to local files. In addition, it provides a preview window for immediate visualization of exported images.
}
