% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qs2.R
\name{w.qread}
\alias{w.qread}
\title{w.qread}
\usage{
w.qread(
  filename,
  return = FALSE,
  envir = base::parent.frame(),
  delete = FALSE,
  nthreads = 1
)
}
\arguments{
\item{filename}{The file name/path.}

\item{return}{Return the saved objects as variables instead of loading them into the target environment. Default: FALSE.

If True, return the object directly when only one object is saved, similar to \code{base::readRDS}; otherwise return a list of all objects.}

\item{envir}{The target environment into which the variables will be loaded. This argument is ignored when \code{return} is \code{TRUE}.}

\item{delete}{Whether to delete the local file after reading is completed. Default: FALSE.}

\item{nthreads}{Consistent with \code{qs2::qs_read}, a single thread is used by default. It is recommended not to use too many cores, as excessive parallelism may increase thread scheduling overhead and lead to negative performance gains.}
}
\value{
NULL.
}
\description{
A wrapper around \code{qs2::qs_read} for loading objects saved with \code{w.qsave} into the R environment.
}
\examples{
x1 <- data.frame(int = sample(1e3, replace=TRUE), num = rnorm(1e3) )
x2 <- data.frame(int = sample(1e4, replace=TRUE), num = rnorm(1e4) )
x3 <- data.frame(int = sample(1e5, replace=TRUE), num = rnorm(1e5) )

#1.entire R environment
w.qsave( 'entire_R_environment.qdata' )
#w.qread( 'entire_R_environment.qdata' )

#2.x1 and x2
w.qsave( 'x1_x2.qdata' , x1, x2 )
#w.qread( 'x1_x2.qdata' )    #load x1 and x2 into the R environment
#x1_x2_list <- w.qread( 'x1_x2.qdata' , return = T )    #return a list object

#3.x1
w.qsave(  'x1.qrds' , x1 )
#w.qread( 'x1.qrds' , return = F )    #load x1 into the R environment
#x1.copy <- w.qread( 'x1.qrds' , return = T )    #assign x1 to a new R object (x1.copy).

}
