% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.markers_decision.R
\name{w.sc.markers_decision}
\alias{w.sc.markers_decision}
\title{Marker decision}
\usage{
w.sc.markers_decision(
  object,
  gmt,
  top = 20,
  level1.threshold = "avg_log2FC > 0.3 & p_val_adj < 0.05 & pct.1 > 0.3"
)
}
\arguments{
\item{object}{Two methods:
\enumerate{
\item Seurat object. A processed Seurat object with computed embeddings (e.g., PCA/UMAP) and defined cluster identities accessible via the \code{Seurat::Idents} function.
\item The raw output returned by this function. It can be provided as input to a subsequent run, in which case the existing \code{DEG} results will be reused for downstream analyses without re-computing differential expression.
}}

\item{gmt}{An absolute file path to a GMT file containing candidate marker genes for each cell type.}

\item{top}{Provide an integer specifying the number of top-ranked positive markers (sorted by avg_log2FC) to retain for downstream analysis.}

\item{level1.threshold}{Differential expression threshold that Level 1 markers must satisfy.}
}
\value{
A list object. The final results are available in the \code{markers} component.

This function comprehensively evaluates the differential expression and expression specificity of marker genes to assign confidence levels. The \code{level} column indicates the confidence level, with level 1 being the highest and level 4 the lowest.

Levels 1 and 2 represent high-confidence markers. Levels 3 and 4 are provided for reference only and should be manually reviewed. Level 1 markers are retained without a strict limit whenever possible. For levels 2–4, no more than ten markers are returned.
}
\description{
Determine the most representative marker genes for each cluster based on the GMT file and differential expression of markers.
}
\examples{

marker1 <- w.sc.markers_decision( object = seurat.obj , gmt = 'test.gmt' , top = 20   )
result1 <- marker1$markers

marker2 <- w.sc.markers_decision( object =  marker1 , gmt = 'test.gmt' , top = 40   )
result2 <- marker2$markers

marker3 <- w.sc.markers_decision( object =  marker2 , gmt = 'new.gmt' , top = 40   )
result3 <- marker3$markers

#result4 produces the same results as result3.
marker4 <- w.sc.markers_decision( object =  seurat.obj , gmt = 'new.gmt' , top = 40   )
result4 <- marker4$markers

}
