% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_matrix.R
\name{wb.convert_id}
\alias{wb.convert_id}
\title{Convert gene IDs}
\usage{
wb.convert_id(
  genes,
  ref,
  ip.type,
  op.type,
  type = "first",
  homologous_species = NULL,
  threads = NULL
)
}
\arguments{
\item{genes}{A vector of gene names.}

\item{ref}{The referenced Bioconductor annotation object typically starts with org and commonly contains the following columns:

ACCNUM,

ALIAS,

ENSEMBL,

ENSEMBLPROT,

ENSEMBLTRANS,

ENTREZID,

ENZYME,

EVIDENCE,

EVIDENCEALL,

GENENAME,

GENETYPE,

GO,

GOALL,

IPI,

MAP,

OMIM,

ONTOLOGY,

ONTOLOGYALL,

PATH,

PFAM,

PMID,

PROSITE,

REFSEQ,

SYMBOL,

UCSCKG,

UNIPROT.}

\item{ip.type}{A vector of one or more gene types. The gene types must exist in the object specified by ref.}

\item{op.type}{Similar to the ip.type, it indicates the gene identifier type to be output.}

\item{type}{Output mode. Options are 'first' or 'any'. When 'first', the result corresponds one-to-one with the input gene vector. When 'any', a single gene may produce multiple rows in the result.}

\item{homologous_species}{A vector of species IDs for homologous genes. Available species IDs can be obtained via homologene::taxData. 9606(Homo sapiens); 10090(Mus musculus).}

\item{threads}{Number of cores used for parallel computation. By default, the maximum computing resources are used}
}
\description{
This function encapsulates the mapIds and select functions from the AnnotationDbi package and the homologene function, providing a unified interface for gene identifier conversion.
}
\examples{
gene_convert <- wb.convert_id( genes = c( 'ENSG00000001167', 'ENSG00000002079', 'ENO2', 'NKG2A'), 
	ref = org.Hs.eg.db, 
	ip.type = c('ENSEMBL', 'ALIAS'), op.type = 'SYMBOL' , homologous_species =  10090 )
}
