% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_normalize.R
\name{wb.data_normalize}
\alias{wb.data_normalize}
\title{Calculate CPM, TPM, and FPKM}
\usage{
wb.data_normalize(count_file, op.dir = ".")
}
\arguments{
\item{count_file}{Absolute path to the raw output file from featureCounts}

\item{op.dir}{Directory to save the result files}
}
\description{
Raw output files from the featureCounts algorithm, used to calculate CPM, TPM, and FPKM, respectively
}
