% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg_pipeline.R
\name{wb.deg_pipeline}
\alias{wb.deg_pipeline}
\title{Perform differential gene expression analysis}
\usage{
wb.deg_pipeline(counts, tpm, group, contrast, p.adj = 0.05, FC = 2, write = F)
}
\arguments{
\item{counts}{Data frame object containing the counts matrix}

\item{tpm}{Data frame object containing the TPM matrix}

\item{group}{Grouping vector for samples (up to two groups)}

\item{contrast}{Comparison vector, e.g., c('M', 'N'). Here, 'N' is the control, and fold change (FC) is calculated as M/N}

\item{p.adj}{P-value threshold}

\item{FC}{Fold change (FC) threshold}

\item{write}{If set to TRUE, the results of differential gene expression analysis from each algorithm will be automatically saved locally}
}
\description{
Automatically calls the DESeq2, edgeR, and limma packages to identify differentially expressed genes
}
