% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_colors.R
\name{wb.ggop}
\alias{wb.ggop}
\title{Previewing and saving ggplot2 objects}
\usage{
wb.ggop(plot, pre = NULL, sur = ".jpg", file = NULL, res = 600, h = 5, w = 5)
}
\arguments{
\item{plot}{ggplot2 object}

\item{pre}{File name prefix}

\item{sur}{File name suffix, default is '.jpg'}

\item{file}{Full file name. If this parameter is provided, pre and sur will be ignored.}

\item{res}{Dots per inch (DPI) resolution}

\item{h}{Height of the image}

\item{w}{Width of the image}
}
\description{
Previewing and saving ggplot2 objects
}
