% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{wb.log_time_runtime}
\alias{wb.log_time_runtime}
\title{Runtime of log}
\usage{
wb.log_time_runtime(t.minor, t.major, prefix = "Runtime: ")
}
\arguments{
\item{t.minor}{Earlier Sys.time object}

\item{t.major}{Later Sys.time object}

\item{prefix}{Prefix}
}
\value{
The returned value is designed to be used with the message function.
}
\description{
Runtime of log
}
