% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{wb.log_time_start_end}
\alias{wb.log_time_start_end}
\title{Started/Completed texts of log}
\usage{
wb.log_time_start_end(format = "s")
}
\arguments{
\item{format}{Options are s (Started) and c (Completed).}
}
\value{
The returned value is designed to be used with the message function.
}
\description{
Started/Completed texts of log
}
