% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{wb.log_time_title}
\alias{wb.log_time_title}
\title{Time of log}
\usage{
wb.log_time_title(format = 1)
}
\arguments{
\item{format}{Format of time. \code{1} (bold and cyan); \code{2} (bold); \code{3} (regular).}
}
\value{
The returned value is designed to be used with the message function.
}
\description{
Time of log
}
