% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc.R
\name{wb.mc}
\alias{wb.mc}
\title{Multithreaded lapply}
\usage{
wb.mc(X, FUN, ..., mc.cores = NULL, pb = T, time = T)
}
\arguments{
\item{X}{Same as the X parameter in \code{wb.smc}.}

\item{FUN}{Same as the FUN parameter in \code{wb.smc}.}

\item{...}{Same as \code{wb.smc}.}

\item{mc.cores}{Number of cores used for parallel computation. By default, the maximum computing resources are used.}

\item{pb}{Show progress bar. Default is TRUE.}

\item{time}{Display execution time. Default is TRUE.}
}
\description{
A special use case of \code{wb.smc}, where memory usage is not limited. By default, all available threads are used for computation.
}
