% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{wb.palette_get}
\alias{wb.palette_get}
\title{Get colors from a palette}
\usage{
wb.palette_get(palette_names = NULL, ncolor = 5, colors = NULL, cat = F)
}
\arguments{
\item{palette_names}{A palette name in SCP package}

\item{ncolor}{Number of colors returned}

\item{colors}{A character vector of color names for previewing. When provided, the palette_names argument will be ignored}

\item{cat}{Whether to print colors}
}
\value{
A color vector.
}
\description{
Get colors from a palette
}
