% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{wb.palette_show_scp}
\alias{wb.palette_show_scp}
\title{Display the palettes in SCP}
\usage{
wb.palette_show_scp(
  palette_names = NULL,
  ncolor = 20,
  ncol = 5,
  h2w = 0.6,
  name.size = 10
)
}
\arguments{
\item{palette_names}{A character vector of palette names in SCP package}

\item{ncolor}{Number of colors to display}

\item{ncol}{Number of columns for displaying colors}

\item{h2w}{The ratio of color bar height to width}

\item{name.size}{Text size for color names}
}
\value{
A ggplot2 object.
}
\description{
Display the palettes in SCP
}
