% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qs2.R
\name{wb.qread}
\alias{wb.qread}
\title{wb.qread}
\usage{
wb.qread(filename, envir = .GlobalEnv, return = F, nthreads = 1)
}
\arguments{
\item{filename}{The file name/path.}

\item{envir}{The target environment into which the variables will be loaded.}

\item{return}{Return the saved objects as variables instead of loading them into the target environment.

If a single object is saved, it is returned directly; if multiple objects are saved, a list containing these objects is returned.}

\item{nthreads}{Consistent with \code{qs2::qs_read}, a single thread is used by default. It is recommended not to use too many cores, as excessive parallelism may increase thread scheduling overhead and lead to negative performance gains.}
}
\value{
NULL.
}
\description{
A wrapper around \code{qs2::qs_read} for loading objects saved with \code{wb.qsave} into the R environment.
}
