% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qs2.R
\name{wb.qsave}
\alias{wb.qsave}
\title{wb.qsave}
\usage{
wb.qsave(
  filename,
  ...,
  envir = .GlobalEnv,
  compress_level = qs2::qopt("compress_level"),
  shuffle = qs2::qopt("shuffle"),
  nthreads = 1
)
}
\arguments{
\item{filename}{The file name/path.

It is recommended to use the \code{.qrds} extension when saving a single object, and \code{.qdata} for all other cases. The suffix has no strict meaning and is only intended to help users choose the appropriate loading method with \code{wb.qread}.}

\item{...}{By default, all objects in the R environment are saved. Alternatively, users can specify multiple objects, in which case only the selected objects will be saved to the local file.}

\item{envir}{The environment in which the objects are located.}

\item{compress_level}{Consistent with the \code{qs2::qs_save}.}

\item{shuffle}{Consistent with the \code{qs2::qs_save}.}

\item{nthreads}{Consistent with the \code{qs2::qs_save}, a single thread is used by default. It is recommended not to use too many cores, as excessive parallelism may increase thread scheduling overhead and lead to negative performance gains.}
}
\value{
NULL.
}
\description{
Wraps the \code{qs2::qs_save} function to enable fast serialization of either the entire R environment or a set of specified objects to a local file, with significantly improved performance compared to the base R functions \code{save}, \code{save.image} and \code{saveRDS}.
}
\examples{
x1 <- data.frame(int = sample(1e3, replace=TRUE), num = rnorm(1e3) )
x2 <- data.frame(int = sample(1e4, replace=TRUE), num = rnorm(1e4) )
x3 <- data.frame(int = sample(1e5, replace=TRUE), num = rnorm(1e5) )

#1.entire R environment
wb.qsave( 'entire_R_environment.qdata' )
#wb.qread( 'entire_R_environment.qdata' )

#2.x1 and x2
wb.qsave( 'x1_x2.qdata' , x1, x2 )
#wb.qread( 'x1_x2.qdata' )

#3.x1
wb.qsave(  'x1.qrds' , x1 )
#wb.qread( 'x1.qrds' , return = F )
#read_x1 <- wb.qread( 'x1.qrds' , return = T )

}
