% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbindlist.R
\name{wb.rbindlist_n}
\alias{wb.rbindlist_n}
\title{Wrapper of rbindlist}
\usage{
wb.rbindlist_n(
  l,
  n = 10000,
  max_chunks = NULL,
  fill = FALSE,
  use.names = TRUE,
  idcol = NULL,
  threads = NULL
)
}
\arguments{
\item{l}{A list containing data.table, data.frame or list objects.}

\item{n}{The maximum number of elements in each sublist. The original list will be split based on this value.}

\item{max_chunks}{The number of chunks to split the original list into. If specified, n is ignored.}

\item{fill}{TRUE fills missing columns with NAs, or NULL for missing list columns. By default FALSE.}

\item{use.names}{TRUE binds by matching column name, FALSE by position.}

\item{idcol}{Creates a column in the result showing which list item those rows came from. TRUE names this column ".id". idcol="file" names this column "file". If the input list has names, those names are the values placed in this id column, otherwise the values are an integer vector 1:length(l).}

\item{threads}{Number of cores used for parallel computation. By default, the maximum computing resources are used.}
}
\value{
An unkeyed data.table containing a concatenation of all the items passed in.
}
\description{
A wrapper around data.table::rbindlist() designed to handle extremely large lists safely by reducing memory pressure and preventing memory overflow.
}
