% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.anno_based_gmt.R
\name{wb.sc.anno_based_gmt}
\alias{wb.sc.anno_based_gmt}
\title{Title}
\usage{
wb.sc.anno_based_gmt(object, gmt, markers, top = 20, write = T, source = "gmt")
}
\arguments{
\item{object}{A processed Seurat object with computed embeddings (e.g., PCA/UMAP) and defined cluster identities accessible via the \code{Seurat::Idents} function.}

\item{gmt}{An absolute file path to a GMT file.}

\item{markers}{The original output results of Seurat's \code{FindAllMarkers} function.}

\item{top}{Provide an integer specifying the number of top-ranked positive markers (sorted by avg_log2FC) to retain for downstream analysis.}

\item{write}{Whether to save the analysis results locally.}

\item{source}{Provide a string representing the local experiment ID. It has no computational significance and is used for reference or labeling purposes only.}
}
\value{
This function returns a data.frame with 10 columns:
\enumerate{
\item cell_type: The annotated cell type.
\item markers: Differentially expressed markers supporting the corresponding \code{cell_type}.
\item markers_number: The number of differentially expressed markers.
\item cluster: The original cluster.
\item cluster_same_as: Clusters that share overlapping differentially expressed markers.
\item not_0_cell_proportion: For the markers in \verb{column 2}, the proportion of non-zero cells in each cluster.
\item not_0_cell_mean_abundance: The mean expression of each marker in non-zero cells.
\item avg_log2FC: The original \code{avg_log2FC} values corresponding to each marker.
\item hit_markers_num_2_all_markers_num: The ratio of \code{markers_number} to the \verb{total number of markers} for the corresponding cell_type.
\item marker_source: For annotation purposes only.
}
}
\description{
This function annotates clusters based on marker gene lists provided in GMT format and the differentially expressed genes of each cluster.
}
