% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.fine_cluster.R
\name{wb.sc.fine_cluster}
\alias{wb.sc.fine_cluster}
\title{Fine cluster}
\usage{
wb.sc.fine_cluster(
  object,
  reduction = "umap",
  dims = NULL,
  downsample = 200,
  seed = 100
)
}
\arguments{
\item{object}{A processed Seurat object with computed embeddings (e.g., PCA/UMAP) and defined cluster identities accessible via the \code{Seurat::Idents} function.}

\item{reduction}{Which dimensionality reduction to use.}

\item{dims}{Dimensions of reduction to use as input. Default to using all dimensions. Custom dimension values can be provided, such as 1:20.}

\item{downsample}{Downsample clusters with a cell count greater than(>) this threshold to the specified value in order to accelerate computation.}

\item{seed}{Random seed used for downsampling.}
}
\value{
A data.frame object.
}
\description{
Reassigns cluster labels in a Seurat object based on inter-cluster distances, such that clusters with higher transcriptomic similarity (i.e., shorter distances) are assigned consecutive numeric labels.

This improves interpretability and facilitates downstream cell type annotation by ensuring that related clusters are grouped together in label space.
}
\examples{
seurat.obj
p1 <- Seurat::DimPlot(  seurat.obj , label = T,repel = T , reduction = "umap" )

# rename
new_name <- wb.sc.fine_cluster( seurat.obj )
new.cluster <- new_name$new_cluster
names(new.cluster) <- new_name$raw_cluster
seurat.obj <- RenameIdents( seurat.obj, new.cluster )
p2 <- Seurat::DimPlot(  seurat.obj , label = T,repel = T , reduction = "umap"  )

p1 | p2



}
