% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.markers_decision.R
\name{wb.sc.markers_decision}
\alias{wb.sc.markers_decision}
\title{Marker decision}
\usage{
wb.sc.markers_decision(object, gmt, top = 30)
}
\arguments{
\item{object}{A processed Seurat object with computed embeddings (e.g., PCA/UMAP) and defined cluster identities accessible via the \code{Seurat::Idents} function.}

\item{gmt}{An absolute file path to a GMT file containing candidate marker genes for each cell type.}

\item{top}{Provide an integer specifying the number of top-ranked positive markers (sorted by avg_log2FC) to retain for downstream analysis.}
}
\value{
A data.frame object. The \code{level} column indicates the confidence level, with level 1 being the highest and level 4 the lowest.
\enumerate{
\item This function comprehensively evaluates the differential expression and expression specificity of marker genes to assign confidence levels. Levels 1 and 2 indicate high confidence, while levels 3 and 4 are for reference only and are recommended to be manually reviewed and filtered. Markers at level 1 should be retained as much as possible. For other levels, return no more than ten markers.
\item Up to five specific marker genes are provided for each cell type.
}
}
\description{
Determine the most representative marker genes for each cluster based on the GMT file and differential expression of markers.
}
