% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolerance_2bp.R
\name{wb.tolerance_merge}
\alias{wb.tolerance_merge}
\title{Merge genomic coordinates}
\usage{
wb.tolerance_merge(data, tolerance = 2)
}
\arguments{
\item{data}{A data frame with four columns representing chromosome, start coordinate, end coordinate, and strand.}

\item{tolerance}{Maximum number of tolerated base pairs (inclusive). Default: 2 bp.}
}
\value{
Data frame.
}
\description{
Merge genomic coordinates allowing a tolerance of up to 2 bp (default).
}
