% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_normalize.R
\name{ww.data_normalize}
\alias{ww.data_normalize}
\title{Calculate CPM, TPM, and FPKM}
\usage{
ww.data_normalize(count_file, op.dir = ".")
}
\arguments{
\item{count_file}{Absolute path to the raw output file from featureCounts}

\item{op.dir}{Directory to save the result files. If set to NULL, the results will not be automatically saved to the local disk.}
}
\description{
Raw output files from the featureCounts algorithm, used to calculate CPM, TPM, and FPKM, respectively
}
\examples{
norm_data <- ww.data_normalize(  'featureCounts.raw_output.txt' , op.fir = '.'  )

}
