% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_colors.R
\name{ww.ggp}
\alias{ww.ggp}
\title{Preview and save images}
\usage{
ww.ggp(
  plot,
  pre = NULL,
  sur = ".png",
  file = NULL,
  res = 600,
  h = 5,
  w = 5,
  mode = "M2"
)
}
\arguments{
\item{plot}{A ggplot2 object or a file path. If a ggplot2 object is provided, it is saved to a local file and then previewed. If a file path is provided, the image is previewed directly.}

\item{pre}{File name prefix.}

\item{sur}{File name suffix, default is '.png'. Supported image formats are identical to those of \link[ggpubr:ggexport]{ggpubr::ggexport}.}

\item{file}{Full file name. If this parameter is provided, pre and sur will be ignored.}

\item{res}{Dots per inch (DPI) resolution.}

\item{h}{Height of the image (inch). For raster plots, the final image height is (h × res) pixels. For vector graphics, the final height is h inches.}

\item{w}{Width of the image (inch). For raster plots, the final image width is (w × res) pixels. For vector graphics, the final width is w inches.}

\item{mode}{There are two preview modes:

(1) M1 shows a preview generated according to the specified parameters. F1 may sometimes differ slightly from the actual saved image, but it can be viewed in a separate graphics window in R.

(2) M2 reloads the exported image from the local file and previews it.

(3) If local image preview is not required, set this parameter to FALSE.}
}
\description{
This function wraps and extends \code{ggpubr::ggexport} to facilitate saving ggplot objects to local files. In addition, it provides a preview window for immediate visualization of exported images.

It can also preview an existing local image by specifying its file path in an interactive GUI.
}
\examples{

###
library(ggpubr)

data("ToothGrowth")
df <- ToothGrowth
df$dose <- as.factor(df$dose)

### ggplot object
bxp <- ggboxplot(df, x = "dose", y = "len", color = "dose", palette = "jco")

#
ww.ggp( plot = bxp , pre = 'bxp'  )
ww.ggp( plot = bxp , pre = 'bxp' , sur = '.pdf'  )
ww.ggp( plot = bxp , file = 'bxp.tiff' )

###local file
ww.ggp(  plot = 'bxp.tiff'  )
#ww.ggp(  plot = 'bxp.pdf'  )


}
