% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{ww.log_time_runtime}
\alias{ww.log_time_runtime}
\title{Runtime of log}
\usage{
ww.log_time_runtime(t.minor, t.major, prefix = "Runtime: ")
}
\arguments{
\item{t.minor}{Earlier Sys.time object}

\item{t.major}{Later Sys.time object}

\item{prefix}{Prefix}
}
\value{
The returned value is designed to be used with the message function.
}
\description{
Runtime of log
}
\examples{
a <- Sys.time()
b <- Sys.time()
diff_time <- ww.log_time_runtime( a , b )

print( diff_time )

}
