% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{ww.log_time_title}
\alias{ww.log_time_title}
\title{Time of log}
\usage{
ww.log_time_title(format = 1)
}
\arguments{
\item{format}{Format of time. \code{1} (bold and cyan); \code{2} (bold); \code{3} (regular).}
}
\value{
The returned value is designed to be used with the \code{message} or \code{cat} function.
}
\description{
Time of log
}
\examples{
message( ww.log_time_title(1) )
message( ww.log_time_title(2) )
message( ww.log_time_title(3) )

}
