% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_matrix.R
\name{ww.mtx_unique_row}
\alias{ww.mtx_unique_row}
\title{Process duplicate row names in the input matrix}
\usage{
ww.mtx_unique_row(exp, raw_row, convert_name, type = "max")
}
\arguments{
\item{exp}{Expression matrix.}

\item{raw_row}{Original gene vector (usually the row names).}

\item{convert_name}{The converted gene vector (corresponding to raw_row) may include NA, and any rows corresponding to NA entries will be excluded from the final converted dataset.}

\item{type}{Method for retaining data. Options are 'max', 'min', 'sum', 'median', or 'mean'. When 'mean' is selected, the mean of multiple rows is returned.}
}
\description{
After ID conversion, multiple rows may correspond to the same ID. This function can be used to process them and retain unique rows.
}
\examples{
mtx_convert <- ww.mtx_unique_row( exp = exp , raw_row = rownames( exp  ),
                                  convert_name  = c( 't1' ,'t2' ,'t3' ,'t2' ,'t4'   )
                                  )


}
