% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc.R
\name{ww.pblapply}
\alias{ww.pblapply}
\title{Lapply with progress bar}
\usage{
ww.pblapply(X, FUN, ..., pb = T, time = T, unlist = F)
}
\arguments{
\item{X}{Same as \link[base:lapply]{base::lapply}.}

\item{FUN}{Same as \link[base:lapply]{base::lapply}.}

\item{...}{Same as \link[base:lapply]{base::lapply}.}

\item{pb}{Show progress bar. Default is TRUE.}

\item{time}{Display execution time. Default is TRUE.}

\item{unlist}{Apply \link[base:unlist]{base::unlist} to the final result. Default is FALSE.}
}
\value{
Returns results consistent with \code{base::lapply}.
}
\description{
A wrapper for \code{base::lapply} that displays a synchronized progress bar.
}
