% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.cluster_clean.R
\name{ww.sc.cluster_clean}
\alias{ww.sc.cluster_clean}
\title{Identify cluster-specific outliers}
\usage{
ww.sc.cluster_clean(
  object,
  group.by,
  reduction = "umap",
  dims = NULL,
  method = "lof",
  minLOF = 2,
  minPts = 20,
  eps = 0.8,
  plot = T
)
}
\arguments{
\item{object}{A processed Seurat object with computed embeddings (e.g., PCA/UMAP).}

\item{group.by}{A column name in \code{object@meta.data} that specifies the group assignment of each cluster.}

\item{reduction}{Which dimensionality reduction to use.}

\item{dims}{Dimensions of reduction to use as input. Default to using all dimensions. Custom dimension values can be provided, such as 1:20.}

\item{method}{Outlier detection can be performed using either \code{lof} or \code{dbscan}, implemented through the \link[dbscan:lof]{dbscan::lof} or \link[dbscan:dbscan]{dbscan::dbscan} functions, respectively.}

\item{minLOF}{The threshold for outlier detection when \code{method = "lof"}. Cells with values above this threshold are considered outliers.}

\item{minPts}{Same as \link[dbscan:lof]{dbscan::lof} or \link[dbscan:dbscan]{dbscan::dbscan}.}

\item{eps}{Same as \link[dbscan:dbscan]{dbscan::dbscan}.}

\item{plot}{Whether to plot the results to evaluate the filtering performance.}
}
\value{
A list object containing the outlier detection results and the corresponding ggplot2 objects.
}
\description{
Detect outlier cells within each cluster based on the low-dimensional embedding matrix, thereby improving the visualization of dimensionality reduction results (e.g., UMAP plots).
}
