% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_5_3.R
\name{ww.seurat_5_3}
\alias{ww.seurat_5_3}
\title{Seurat object version conversion}
\usage{
ww.seurat_5_3(seurat, v = "V3")
}
\arguments{
\item{seurat}{Seurat object}

\item{v}{Version of the assay output. Optional V5 and V3, default is 'V3'.}
}
\description{
Seurat object version conversion
}
\examples{
testsc <- chenlab::chenlab_testsc
testsc_v3 <- ww.seurat_5_3( testsc  )

}
