% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc.R
\name{ww.smc}
\alias{ww.smc}
\title{Smart mclapply}
\usage{
ww.smc(
  X,
  FUN,
  ...,
  mc.cores = NULL,
  mem.ratio.max = 0.8,
  mem.max = NULL,
  pb = T,
  time = T,
  unlist = F
)
}
\arguments{
\item{X}{Same as the X parameter in \link[parallel:mcdummies]{parallel::mclapply}.}

\item{FUN}{Same as the FUN parameter in \link[parallel:mcdummies]{parallel::mclapply}.}

\item{...}{Same as \link[parallel:mcdummies]{parallel::mclapply}.}

\item{mc.cores}{Maximum number of cores used for parallel computation. By default, the number of cores is automatically determined based on memory usage, ensuring that memory does not overflow while fully utilizing all available system resources.

If an integer is provided, the program will force execution with the specified number of threads, ignoring memory protection.}

\item{mem.ratio.max}{Maximum proportion of available memory allowed when automatically determining the number of cores.}

\item{mem.max}{Maximum amount of memory (in GB) allowed when automatically determining the number of cores. Defaults to NULL, which will automatically detect the maximum available system memory.}

\item{pb}{Show progress bar. Default is TRUE.}

\item{time}{Display execution time. Default is TRUE.}

\item{unlist}{Apply \link[base:unlist]{base::unlist} to the final result. Default is FALSE.}
}
\value{
Returns results consistent with \code{parallel::mclapply}.
}
\description{
A smart mclapply function that automatically determines the number of cores for parallel computation, maximizing the use of system resources while ensuring memory does not overflow.

Its core logic is to split the input vector into chunks for execution, thereby preventing memory overflow.
}
