% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_fair.R
\name{ww.split_fair}
\alias{ww.split_fair}
\title{Fair split}
\usage{
ww.split_fair(x, chunk.length = NULL, number = NULL, min.length = 2)
}
\arguments{
\item{x}{A vector / list / factor to be split.}

\item{chunk.length}{Target length of each subset. If provided, the input vector will be split into chunks with approximately this length.}

\item{number}{Target number of subsets. If provided, the input vector will be evenly split into this many parts.}

\item{min.length}{Minimum allowed length of each subset. If the requested split would produce subsets smaller than this value, the number of subsets will be reduced automatically to satisfy this constraint.}
}
\value{
A list of subsets.
}
\description{
Split a vector into approximately equal-sized subsets, either by specifying the target chunk length or the number of subsets, while enforcing a minimum subset size.
}
\examples{
a1 <- ww.split_fair( 1:21 , chunk.length = 5 , min.length = 2  )
a2 <- ww.split_fair( 1:21 , chunk.length = 5 , min.length = 1  )

a3 <- ww.split_fair( 1:21 , number = 6 , min.length = 2  )

}
