% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_matrix.R
\name{wb.convert_id}
\alias{wb.convert_id}
\title{Convert gene IDs}
\usage{
wb.convert_id(
  genes,
  ref,
  ip.type = NULL,
  op.type,
  type = "first",
  homologous_species = NULL,
  threads = NULL
)
}
\arguments{
\item{genes}{A vector of gene names.}

\item{ref}{The referenced Bioconductor annotation object typically starts with org and commonly contains the following columns:

ACCNUM,

ALIAS,

ENSEMBL,

ENSEMBLPROT,

ENSEMBLTRANS,

ENTREZID,

ENZYME,

EVIDENCE,

EVIDENCEALL,

GENENAME,

GENETYPE,

GO,

GOALL,

IPI,

MAP,

OMIM,

ONTOLOGY,

ONTOLOGYALL,

PATH,

PFAM,

PMID,

PROSITE,

REFSEQ,

SYMBOL,

UCSCKG,

UNIPROT.}

\item{ip.type}{A vector of one or more gene types. Gene types must exist in the object specified by ref. Defaults to NULL, which enables automatic detection.}

\item{op.type}{Similar to the ip.type, it indicates the gene identifier type to be output.}

\item{type}{Output mode. Options are 'first' or 'any'. When set to "first" and the length of op.type is 1, the result corresponds one-to-one with the input gene vector. When 'any', a single gene may produce multiple rows in the result.}

\item{homologous_species}{A vector of species IDs for homology conversion. If specified, homologous gene IDs from the specified species will also be returned after gene ID conversion. Available species IDs can be obtained via homologene::taxData, e.g. 9606 (Homo sapiens), 10090 (Mus musculus).}

\item{threads}{Number of cores used for parallel computation. By default, the maximum computing resources are used}
}
\description{
This function encapsulates the mapIds and select functions from the AnnotationDbi package and the homologene function, providing a unified interface for gene identifier conversion.
}
\examples{
gene_convert <- wb.convert_id( genes = c( 'ENSG00000001167', 'ENSG00000002079', 'ENO2', 'NKG2A'),
	ref = org.Hs.eg.db,
	ip.type = c('ENSEMBL', 'ALIAS'), op.type = c('SYMBOL', 'ENTREZID'),
	homologous_species =  10090 )
}
