% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_matrix.R
\name{wb.convert_id}
\alias{wb.convert_id}
\title{Convert gene IDs}
\usage{
wb.convert_id(genes, ref, ip.type, op.type, type = "first", mc.cores = NULL)
}
\arguments{
\item{genes}{Names of the gene vector}

\item{ref}{Reference org annotation object}

\item{ip.type}{A vector containing one or more types can be provided}

\item{op.type}{A vector containing one or more types can be provided}

\item{type}{Output mode. Options are 'first' or 'any'. When 'first', the result corresponds one-to-one with the input gene vector. When 'any', a single gene may produce multiple rows in the result.}

\item{mc.cores}{Number of cores used for parallel computation. By default, the maximum computing resources are used}
}
\description{
This function is a wrapper around mapIds and select from AnnotationDbi, used for converting gene IDs.
}
