% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_colors.R
\name{ww.ggp}
\alias{ww.ggp}
\title{Preview and save images}
\usage{
ww.ggp(
  plot,
  pre = NULL,
  sur = ".png",
  file = NULL,
  res = 600,
  h = 5,
  w = 5,
  mode = "M2"
)
}
\arguments{
\item{plot}{File path, base or grid plot, or graphic generated by ggplot, lattice, etc.

If a file path is provided, use it directly for previewing. Otherwise, save the file locally first and then preview it. Input:
\itemize{
\item An image file path.
\item A ggplot2 object.
\item A \code{{}} expression containing plotting code.
}}

\item{pre}{File name prefix.}

\item{sur}{File name suffix, default is '.png'. Supported image formats are identical to those of \link[ggpubr:ggexport]{ggpubr::ggexport}.}

\item{file}{Full file name. If this parameter is provided, pre and sur will be ignored.}

\item{res}{Dots per inch (DPI) resolution.}

\item{h}{Height of the image (inch). For raster plots, the final image height is (h × res) pixels. For vector graphics, the final height is h inches.}

\item{w}{Width of the image (inch). For raster plots, the final image width is (w × res) pixels. For vector graphics, the final width is w inches.}

\item{mode}{There are two preview modes:
\itemize{
\item M1. M1 shows a preview generated according to the specified parameters. F1 may sometimes differ slightly from the actual saved image, but it can be viewed in a separate graphics window in R.
\item M2. M2 reloads the exported image from the local file and previews it.
\item FALSE. If local image preview is not required, set this parameter to FALSE.
}}
}
\description{
This function wraps and extends \code{ggpubr::ggexport} to facilitate saving ggplot objects to local files. In addition, it provides a preview window for immediate visualization of exported images.

It can also preview an existing local image by specifying its file path in an interactive GUI.
}
\examples{

###
library(ggpubr)

data("ToothGrowth")
df <- ToothGrowth
df$dose <- as.factor(df$dose)

### 1. ggplot object
bxp <- ggboxplot(df, x = "dose", y = "len", color = "dose", palette = "jco")

#
ww.ggp( plot = bxp , pre = 'bxp'  )
ww.ggp( plot = bxp , pre = 'bxp' , sur = '.pdf'  )
ww.ggp( plot = bxp , file = 'bxp.tiff' )

### 2. grid plot
#bar
ww.ggp( plot = {
  data = 1:10
  barplot(data)
}, pre = 'bxp.grid' )

#heatmap
heatmap.data <- matrix( sample(1:1000 ,30) , nrow =  3 )
ww.ggp( plot = {
  pheatmap::pheatmap( heatmap.data )
} , pre = 'heatmap.grid' , h = 4 , w = 8  )

### 3. local file
ww.ggp(  plot = 'bxp.tiff'  )
#ww.ggp(  plot = 'bxp.pdf'  )


}
