% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_colors.R
\name{ww.group_colors}
\alias{ww.group_colors}
\title{Obtain multiple sets of colors}
\usage{
ww.group_colors(number, continuous = F)
}
\arguments{
\item{number}{Number of colors to return}

\item{continuous}{Whether to return a continuous color scale}
}
\description{
Up to 50 groups
}
