% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_manage.R
\name{ww.package_install}
\alias{ww.package_install}
\title{Install package}
\usage{
ww.package_install(
  package.name,
  method = "I",
  interactive.install = TRUE,
  check.only = FALSE
)
}
\arguments{
\item{package.name}{Package name.}

\item{method}{Method used for automatic installation. Can be "I", "B", or a character string.
\enumerate{
\item If "I", the package will be installed using utils::install.packages().
\item If "B", the package will be installed using BiocManager::install().
\item A custom installation command can also be provided as a character string, e.g. "utils::install.packages('dplyr')".
}}

\item{interactive.install}{Whether to provide an interactive prompt asking the user to confirm before starting the installation.}

\item{check.only}{Whether to perform package checking only without automatic installation.}
}
\value{
\enumerate{
\item TRUE: Returned when the package check passes or the package is installed successfully.
\item FALSE: Returned when the package check fails or the installation is unsuccessful.
}
}
\description{
Check whether a package is installed and optionally install it automatically using a specified installation method.
}
