% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{ww.palette_view}
\alias{ww.palette_view}
\title{Get colors from a SCP palette / Preview palette}
\usage{
ww.palette_view(palette_names = NULL, ncolor = 5, colors = NULL, cat = F)
}
\arguments{
\item{palette_names}{A palette name in SCP package}

\item{ncolor}{Number of colors returned}

\item{colors}{A character vector of color names for previewing. When provided, the palette_names argument will be ignored}

\item{cat}{Whether to print colors}
}
\value{
A named vector of color names.
}
\description{
Get colors from a SCP palette / Preview palette
}
