% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qs2.R
\name{ww.qsave}
\alias{ww.qsave}
\title{ww.qsave}
\usage{
ww.qsave(
  filename,
  ...,
  force = F,
  envir = base::parent.frame(),
  version = 2,
  compress_level = qs2::qopt("compress_level"),
  shuffle = qs2::qopt("shuffle"),
  nthreads = 1
)
}
\arguments{
\item{filename}{The file name/path. The file name does not affect the behavior of the function.

We recommend using the \code{.qrds} extension when saving a single object, and \code{.qdata} for all other cases. The suffix has no strict meaning and is only intended to help users choose the appropriate loading method with \link{ww.qread}.}

\item{...}{By default, all objects in the R environment are saved. Alternatively, users can specify multiple objects, in which case only the selected objects will be saved to the local file.}

\item{force}{Whether to force overwriting an existing file. Default is \code{FALSE}. If \code{filename} already exists, the file will not be overwritten and a message will be displayed.

Set to \code{TRUE} to overwrite the existing file.}

\item{envir}{The environment in which the objects are located.}

\item{version}{Version of the serialization package to use. Integer value:
\itemize{
\item 1: use the \code{qs} package
\item 2: use the \code{qs2} package
}}

\item{compress_level}{As with \link[qs:qsave]{qs::qsave} or \link[qs2:qs_save]{qs2::qs_save}.}

\item{shuffle}{As with the \link[qs2:qs_save]{qs2::qs_save}.}

\item{nthreads}{As with \link[qs:qsave]{qs::qsave} or \link[qs2:qs_save]{qs2::qs_save}, a single thread is used by default. It is recommended not to use too many cores, as excessive parallelism may increase thread scheduling overhead and lead to negative performance gains.}
}
\value{
After successfully saving, return the name of the file saved locally.
}
\description{
Wraps the \code{save} function across different versions of the \code{qs} package to enable fast serialization of either the entire R environment or a set of specified objects to a local file, with significantly improved performance compared to the base R functions \code{save}, \code{save.image} and \code{saveRDS}.
}
\examples{
x1 <- data.frame(int = sample(1e3, replace=TRUE), num = rnorm(1e3) )
x2 <- data.frame(int = sample(1e4, replace=TRUE), num = rnorm(1e4) )
x3 <- data.frame(int = sample(1e5, replace=TRUE), num = rnorm(1e5) )

#1. entire R environment
ww.qsave( 'entire_R_environment.qdata' )
# ww.qread( 'entire_R_environment.qdata' )

#2. x1 and x2
ww.qsave( 'x1_x2.qdata' , x1, x2 )
# ww.qread( 'x1_x2.qdata' )    #load x1 and x2 into the R environment
# x1_x2_list <- ww.qread( 'x1_x2.qdata' , return = T )    #return a list object

#3. x1
ww.qsave(  'x1.qrds' , x1 )
# ww.qread( 'x1.qrds' , return = F )    #load x1 into the R environment
# x1.copy <- ww.qread( 'x1.qrds' , return = T )    #assign x1 to a new R object (x1.copy).

}
