% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc.metacell.R
\name{ww.sc.metacell}
\alias{ww.sc.metacell}
\title{Metacell for seurat object}
\usage{
ww.sc.metacell(
  object,
  cluster,
  group.by = NULL,
  assay = NULL,
  layer = NULL,
  method = "mine",
  min.cluster = 50,
  min.cluster.reserved = T,
  min.target = 50,
  max.target = NULL
)
}
\arguments{
\item{object}{A processed Seurat object.}

\item{cluster}{The column name in the Seurat object's meta.data that indicates cell types.}

\item{group.by}{The column name in the Seurat object's meta.data that indicates the group. If specified, the object will be split by this column prior to metacell generation, creating metacells independently for each subset.}

\item{assay}{Same as \link[Seurat:reexports]{Seurat::GetAssayData}.}

\item{layer}{Same as \link[Seurat:reexports]{Seurat::GetAssayData}.}

\item{method}{Metacell generation algorithms:
\itemize{
\item mine: using our algorithm, we first hierarchically clustered all single cells and then aggregated highly similar cells into metacells. The proportional scaling strategy is used, with the maximum number of metacells determined by the max.target parameter.
For large-scale datasets, this method achieves higher computational efficiency than SuperCell and other existing approaches, with a more substantial improvement when the \code{max.target} parameter is applied.
\item supercell: use the \code{SuperCell} package with the same scale factor.
\item hdwgcna: use the \code{hdWGCNA} package
}}

\item{min.cluster}{Clusters containing more cells than this value are used for metacell generation. Recommendation: 50.

The value of \code{min.cluster} cannot be smaller than \code{min.target}!!!}

\item{min.cluster.reserved}{Whether to retain clusters containing fewer cells than min.cluster. When TRUE, the original expression values of these clusters are retained rather than the metacell values. Clusters with only 1 cell are always excluded.}

\item{min.target}{The expected number of cells for the smallest cluster after metacell generation. This value is used to calculate the scaling factor. If the smallest cluster contains an actual number of N cells, the scaling factor is calculated as N / min.target. All other clusters are scaled according to this factor. Recommendation: 50.}

\item{max.target}{Available only when method = "mine". This parameter defines the maximum number of cells in the largest cluster after proportional scaling. Clusters larger than this limit will no longer be scaled proportionally.}
}
\value{
A list containing: (1) the metacell expression matrix (averaged), and (2) the cluster assignment of each metacell.
}
\description{
Generates a metacell expression matrix for a Seurat object, where each metacell is constructed by aggregating multiple single cells and its expression is defined as the mean expression across the constituent cells.
}
